requirements.txt
# --- Utilities & Security ---
python-dotenv==1.0.1
requests==2.31.0
tenacity==8.2.3
boto3==1.34.0 # Thư viện kết nối MinIO/S3 cho Batch Path

# --- Data Processing ---
pyspark==3.5.0 # Thư viện cho Apache Spark
pandas==2.2.1
duckdb==1.1.3 # Nâng cấp lên bản V1 stable mới nhất

# --- Streaming (Kafka) ---
confluent-kafka==2.3.0

# --- Time-series Database ---
influxdb-client==1.36.0